Code refactor + update tests
